refactor: renamed bundle for configurable component
