+++ kura-addon-prototypes/bundles/example-configurable-component.bundle/build.properties
