--- kura-addon-prototypes/bundles/example-simple-component.bundle/META-INF/MANIFEST.MF
@@ -1,14 +1,0 @@
-Manifest-Version: 1.0
-Bundle-ManifestVersion: 2
-Bundle-SymbolicName: org.eclipse.kura.example.simple.component.bundle;singleton:=true
-Bundle-Version: 1.0.0.qualifier
-Bundle-Vendor: Eclipse Kura
-Require-Capability: osgi.ee;filter:="(&(osgi.ee=JavaSE)(version=17))"
-Import-Package: org.eclipse.kura;version="[1.4,2.0)",
- org.eclipse.kura.configuration;version="[1.0,2.0)",
- org.eclipse.kura.configuration.metatype;version="[1.1,2.0)",
- org.eclipse.kura.core.configuration;version="[2.0,3.0)",
- org.eclipse.kura.core.configuration.metatype;version="[1.0,2.0)",
- org.slf4j;version="1.7"
-Bundle-ActivationPolicy: lazy
-Service-Component: OSGI-INF/*.xml
