--- kura-addon-prototypes/bundles/example-simple-component.bundle/pom.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Copyright (c) 2025 Eurotech and/or its affiliates and others
-
-    This program and the accompanying materials are made
-    available under the terms of the Eclipse Public License 2.0
-    which is available at https://www.eclipse.org/legal/epl-2.0/
-
-    SPDX-License-Identifier: EPL-2.0
-
-    Contributors:
-     Eurotech
-
--->
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>org.eclipse.kura</groupId>
-        <artifactId>kura-addon-prototypes.parent</artifactId>
-        <version>6.0.0-SNAPSHOT</version>
-    </parent>
-
-    <properties>
-        <sonar.coverage.jacoco.xmlReportPaths>
-            ${project.basedir}/../../tests/*/target/site/jacoco-aggregate/jacoco.xml
-        </sonar.coverage.jacoco.xmlReportPaths>
-    </properties>
-
-    <artifactId>org.eclipse.kura.example.simple.component.bundle</artifactId>
-    <version>1.0.0-SNAPSHOT</version>
-    <packaging>eclipse-plugin</packaging>
-
-</project>
