+++ kura-addon-prototypes/bundles/pom.xml
@@ -343,6 +343,6 @@
 
     <modules>
         <module>example-driver.bundle</module>
-        <module>example-simple-component.bundle</module>
+        <module>example-configurable-component.bundle</module>
     </modules>
 </project>
