+++ kura-addon-prototypes/tests/example-configurable-component.bundle.test/build.properties
