--- kura-addon-prototypes/tests/example-simple-component.bundle.test/META-INF/MANIFEST.MF
@@ -1,17 +1,0 @@
-Manifest-Version: 1.0
-Bundle-ManifestVersion: 2
-Bundle-Name: Example component tests
-Bundle-SymbolicName: org.eclipse.kura.example.simple.component.bundle.test;singleton:=true
-Bundle-Version: 1.0.0.qualifier
-Bundle-Vendor: EUROTECH
-Require-Capability: osgi.ee;filter:="(&(osgi.ee=JavaSE)(version=17))"
-Fragment-Host: org.eclipse.kura.example.simple.component.bundle
-Require-Bundle: org.junit,
- org.mockito.mockito-core,
- net.bytebuddy.byte-buddy,
- net.bytebuddy.byte-buddy-agent,
- org.objenesis
-Import-Package: org.eclipse.kura;version="1.5.0",
- org.eclipse.kura.configuration;version="1.1.2",
- org.slf4j;version="1.7.25"
-Service-Component: OSGI-INF/*.xml
