+++ kura-addon-prototypes/tests/pom.xml
@@ -388,7 +388,7 @@
 
     <modules>
         <module>example-driver.bundle.test</module>
-        <module>example-simple-component.bundle.test</module>
+        <module>example-configurable-component.bundle.test</module>
     </modules>
 
 </project>
